Added SSH session checks
-Check if we are in an SSH session in both .profile and setup scripts
-If we are in an SSH session, skip install of nerdfonts

Closes Github #7

+++ __setup/setup.sh
@@ -40,14 +40,17 @@ mkdir ~/tmp ~/log ~/src ~/lib
 
 
 # Install Nerdfonts
-pushd ~/src
-REPO=nerd-fonts
-git clone https://github.com/ryanoasis/$REPO.git --depth 1
-pushd ./$REPO
-./install.sh
-popd
-rm -rf ./$REPO
-popd
+# We don't need to install if we are in an SSH session.
+if [[ $SESSION_TYPE != "remote/ssh" ]]; then
+    pushd ~/src
+    REPO=nerd-fonts
+    git clone https://github.com/ryanoasis/$REPO.git --depth 1
+    pushd ./$REPO
+    ./install.sh
+    popd
+    rm -rf ./$REPO
+    popd
+fi
 
 
 # Flag that we have done initial setup
